Give more flexibility to translators

+++ includes/class-bc-setup.php
@@ -400,7 +400,7 @@ class BC_Setup {
 		$player_api = new BC_Player_Management_API2();
 		$players = $player_api->get_all_players();
 
-		if ( count( $bc_accounts->get_sanitized_all_accounts() ) > 0 && ! empty( $players ) ) {
+		if ( count( $bc_accounts->get_sanitized_all_accounts() ) > 0 && ! empty( $players ) && is_array( $players ) ) {
 
 			if ( false !== get_option( '_brightcove_plugin_activated' ) ) {
 				delete_option( '_brightcove_plugin_activated' );
@@ -410,13 +410,12 @@ class BC_Setup {
 
 		}
 
-		if ( count( $bc_accounts->get_sanitized_all_accounts() ) > 0 && empty( $players ) ) {
+		if ( count( $bc_accounts->get_sanitized_all_accounts() ) > 0 && empty( $players ) && is_array( $players ) ) {
 			$notices[] = array(
 				'message' => sprintf(
-					'%s <a href="%s"><strong>%s</strong></a>',
-					esc_html__( 'It looks like one or more of your accounts API authentication changed recently. Please update your settings ', 'brightcove' ),
-					esc_url( admin_url( 'admin.php?page=brightcove-sources' ) ),
-					esc_html__( 'here', 'brightcove' )
+				/* translators: %s: Settings Screen URL. */
+					__( 'It looks like one or more of your accounts API authentication changed recently. Please update your settings <a href="%s"><strong>here</strong></a>', 'brightcove' ),
+					esc_url( admin_url( 'admin.php?page=brightcove-sources' ) )
 				),
 				'type'    => 'error',
 			);
